Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-RESTx
Flask-Swagger-UI
psycopg2-binary
pycryptodome
python-dotenv
